backend: moved AsJst from streak.rs into utils.rs
